refactor(tools): move TradeContext pooling to a separate PR (#135)
Splits out everything pooling-related from this PR to keep it
independently reviewable: reverts src/tools/trade.rs, src/tools/portfolio.rs,
src/main.rs, src/metrics.rs, and src/ws_pool.rs to their pre-pooling
state, removes src/context_pool.rs and src/trade_pool.rs, and drops
McpContext::get_trade_context()/evict_trade_context() plus the
trade-pooling compile-time lint from src/tools/mod.rs.

The pooling work itself now lives on feat/trade-context-pooling (#135),
unchanged from what was here. This PR is left with: the option_quote
symbol-format documentation fix, the candlesticks count validation/
boundary-retry fix, the error_hint structured-error-code improvements,
and the version bump.
