+++ src/main.rs
@@ -1,7 +1,6 @@
 #![recursion_limit = "256"]
 
 mod auth;
-mod context_pool;
 mod counter;
 mod error;
 mod logging;
@@ -10,7 +9,6 @@ mod serialize;
 #[cfg(test)]
 mod test_support;
 mod tools;
-mod trade_pool;
 mod ws_pool;
 
 use std::net::SocketAddr;
